+++ system/modules/isotope/languages/en/tl_iso_orderstatus.php
@@ -32,6 +32,7 @@
  */
 $GLOBALS['TL_LANG']['tl_iso_orderstatus']['name']			= array('Name', 'Please enter a name for this status.');
 $GLOBALS['TL_LANG']['tl_iso_orderstatus']['paid']			= array('Order is paid', 'Assume the order is paid when it has this status. This will for example allow files to be downloaded.');
+$GLOBALS['TL_LANG']['tl_iso_orderstatus']['welcomescreen']	= array('Show on welcome screen', 'Show number of orders with this status on the backend welcome screen.');
 $GLOBALS['TL_LANG']['tl_iso_orderstatus']['mail_customer']	= array('E-Mail to customer', 'Select an email template to notify the customer when this status is assigned to an order.');
 $GLOBALS['TL_LANG']['tl_iso_orderstatus']['mail_admin']		= array('E-Mail to admin', 'Select an email template to notify the admin when this status is assigned to an order.');